+++ lib/stannum/constraints/parameters.rb
@@ -5,6 +5,7 @@ require 'stannum/constraints'
 module Stannum::Constraints
   # Namespace for constraints that match method parameters.
   module Parameters
-    autoload :ExtraKeywords, 'stannum/constraints/parameters/extra_keywords'
+    autoload :ExtraArguments, 'stannum/constraints/parameters/extra_arguments'
+    autoload :ExtraKeywords,  'stannum/constraints/parameters/extra_keywords'
   end
 end